+++ Examples/examples_bar/bartutex6.php
@@ -1,11 +1,15 @@
-<?php // content="text/plain; charset=utf-8"
+<?php
+
+/**
+ * JPGraph v3.6.15
+ */
 require_once '../../vendor/autoload.php';
 use Amenadiel\JpGraph\Graph;
 use Amenadiel\JpGraph\Plot;
 
 // Some data
 $months = $gDateLocale->GetShortMonth();
-srand((double) microtime() * 1000000);
+srand((float) microtime() * 1000000);
 for ($i = 0; $i < 25; ++$i) {
     $databary[] = rand(1, 50);
     $databarx[] = $months[$i % 12];
@@ -16,7 +20,7 @@ $graph = new Graph\Graph(300, 200, 'auto');
 $graph->SetShadow();
 
 // Use a "text" X-scale
-$graph->SetScale("textlin");
+$graph->SetScale('textlin');
 
 // Specify X-labels
 $graph->xaxis->SetTickLabels($databarx);
@@ -26,14 +30,14 @@ $graph->xaxis->SetTextLabelInterval(3);
 $graph->xaxis->HideTicks();
 
 // Set title and subtitle
-$graph->title->Set("Bar tutorial example 6");
+$graph->title->Set('Bar tutorial example 6');
 
 // Use built in font
 $graph->title->SetFont(FF_FONT1, FS_BOLD);
 
 // Create the bar plot
 $b1 = new Plot\BarPlot($databary);
-$b1->SetLegend("Temperature");
+$b1->SetLegend('Temperature');
 $b1->SetWidth(0.4);
 
 // The order the plots are added determines who's ontop
